Fixed rail_fence.py implementation
After numerious pdb.set_trace() the problem lied in the secret_messages.py decryption function. I was calling for the classes encryption function instead of the decryption. It did not matter with Atbash because it is the same code regardless of encryption/decryption. I adjusted atbash.py's decryption method to call on the encryption method as a result.