One time key encryption fully implemented
The fix was in the one_time_key_encryption() method and was not related to secret_messages.py.  I was increasing the Index_counter by two on the else: by accident. This caused rare bug where if two high numbers were added together in the combined_numbers variable, and if that index was skipped by the portion that reduces the high combined_numbers portion. This allowed for an IndexError latter down the line.