Secret_messages V 1.4 Caesar update
I was adding the Caesar cipher and figured I would update the project. I have left as much as I could from the old project in place to preserve how the project used to look. As a side note. The project seems to have implemented classes well. All I need to do to add a cipher is an import and check if the user wanted the Ceasar cipher.